+++ src/main/java/ru/murlov/dto/CurrencyResponse.java
@@ -1,7 +1,7 @@
 package ru.murlov.dto;
 
-public record CurrencyResponse(Long id,
-                               String code,
-                               String name,
-                               String sign){
+public record CurrencyResponse (int id,
+                                String code,
+                                String name,
+                                String sign) {
 }
